Moved layout components to dedicated frontend folder
