+++ apps/frontend/src/layout/navbar/NavBar.tsx
