+++ apps/frontend/src/layout/pages/Home.css
