+++ apps/frontend/src/layout/pages/ShowListPage.tsx
@@ -1,4 +1,4 @@
-import { ShowList } from "../components/showlist/ShowList"
+import { ShowList } from "../../components/showlist/ShowList"
 
 
 export const ShowListPage = () => (
