+++ apps/frontend/src/layout/pages/WatchProgressPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
-import { useWatchProgress } from "../components/watchprogress/WatchProgress";
+import { useWatchProgress } from "../../components/watchprogress/WatchProgress";
 import "../components/watchprogress/WatchProgress.css";
-import type { WatchProgress } from "../components/types/WatchProgressType";
+import type { WatchProgress } from "../../../../../shared/types/WatchProgressType";
 
 export const WatchProgressPage: React.FC = () => {
   const {
